+++ NEWS.md
@@ -18,6 +18,16 @@ The latest interface to optimizers in tensorflow are now used, these changes are
 * adds `nadam` optimiser - see [docs](https://www.tensorflow.org/api_docs/python/tf/keras/optimizers/Nadam).
 * In `rms_prop` optimiser, changes `decay` parameter to `rho`, and adds `centered` parameter - see [docs](https://www.tensorflow.org/api_docs/python/tf/keras/optimizers/RMSprop)
 
+The following optimisers are removed, as they are no longer supported by Tensorflow:
+
+ * `powell()`
+ * `cg()`
+ * `newton_cg()`
+ * `l_bfgs_b()`
+ * `tnc()`
+ * `cobyla()`
+ * `slsqp()`
+
 ## Minor
 
 * `greta_sitrep()` now checks for minimum versions of software, instead of exact versions. It requires at least Python version 3.8, TensorFlow 2.8.0, and Tensorflow Probability 0.14.0.